digital_code_automating/requirements.txt
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
openpyxl
mysql-connector-python